refactor(lookup): extraire AbstractGeminiLookupProvider pour mutualiser le code Gemini
Mutualise callGemini, consumeRateLimit, parseJsonFromText et
prepareWithCache entre GeminiLookup et BedethequeLookup.
Ajoute publishedDate, tomeNumber et tomeEnd au provider Bedetheque.

#119
